+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- The associated configuration is now logged for each issued secret ([#413]).
+
 ### Changed
 
 - [BREAKING] The TLS CA Secret is now installed into the Namespace of the operator (typically `stackable-operators`), rather than `default` ([#397]).
@@ -19,6 +23,7 @@ All notable changes to this project will be documented in this file.
 
 [#397]: https://github.com/stackabletech/secret-operator/pull/397
 [#403]: https://github.com/stackabletech/secret-operator/pull/403
+[#413]: https://github.com/stackabletech/secret-operator/pull/413
 [#440]: https://github.com/stackabletech/secret-operator/pull/440
 
 ## [24.3.0] - 2024-03-20